# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Auth-All",
    "version": "1.0.0"
  },
  "paths": {
    "/api/auth/account/link/{provider}": {
      "post": {
        "operationId": "accountLink",
        "summary": "Start an explicit provider link",
        "tags": [
          "account"
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "description": "The configured provider id, for example github.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The provider authorization URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkResponse"
                }
              }
            }
          },
          "401": {
            "description": "Auth-All error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Auth-All error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-authall-client": {
          "namespace": "account",
          "method": "link"
        }
      }
    },
    "/api/auth/account/providers": {
      "get": {
        "operationId": "accountProviders",
        "summary": "List the authentication methods of the current user",
        "tags": [
          "account"
        ],
        "responses": {
          "200": {
            "description": "The linked providers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvidersResponse"
                }
              }
            }
          },
          "401": {
            "description": "Auth-All error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-authall-client": {
          "namespace": "account",
          "method": "providers"
        }
      }
    },
    "/api/auth/account/unlink/{provider}": {
      "post": {
        "operationId": "accountUnlink",
        "summary": "Remove a linked provider",
        "tags": [
          "account"
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "description": "The configured provider id, for example github.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The provider is unlinked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Auth-All error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Auth-All error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Auth-All error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-authall-client": {
          "namespace": "account",
          "method": "unlink"
        }
      }
    },
...
    "/api/auth/oauth/{provider}": {
      "get": {
        "operationId": "oauthAuthorize",
        "summary": "Start a provider sign-in",
        "tags": [
          "oauth"
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "description": "The configured provider id, for example github.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "redirect_to",
            "in": "query",
            "description": "Where to send the browser after success.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "A redirect to the provider"
          },
          "404": {
            "description": "Auth-All error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-authall-client": {
          "namespace": "oauth",
          "method": "authorize",
          "redirect": true
        }
      }
    },
    "/api/auth/oauth/{provider}/callback": {
      "get": {
        "operationId": "oauthCallback",
        "summary": "Complete a provider sign-in",
        "tags": [
          "oauth"
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "description": "The configured provider id, for example github.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "A redirect back to the application"
          },
          "400": {
            "description": "Auth-All error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Auth-All error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Auth-All error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-authall-client": {
          "namespace": "oauth",
          "method": "callback",
          "redirect": true
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "LinkResponse": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        },
        "required": [
          "url"
        ]
      },
...
      "ProvidersResponse": {
        "type": "object",
        "properties": {
          "hasPassword": {
            "type": "boolean"
          },
          "providers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "accountId": {
                  "type": "string"
                },
                "linkedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "provider": {
                  "type": "string"
                }
              },
              "required": [
                "provider",
                "accountId",
                "linkedAt"
              ]
            }
          }
        },
        "required": [
          "providers",
          "hasPassword"
        ]
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
...
    }
  }
}